Moved mod partition to a seperate file.
